+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The `sketch_plot` option is now implemented.
+- 'ignore_unconnected' preflight wasn't working.
 
 ## [0.4.0] - 2020-06-17
 ### Added